feat(token verification): Now user can directly borrow or return books. when a user send request to borrow or return book, we use the info in token to validate user
Before we should have provided username of user to borrow or return book, but now we have provided RBAC so that when a user(onced logged in) send a request to borrow or return endpoint, we validate the token in header to get user info and add or return book from that user only. On admin side, admin now can provide username same as before to add or return book on behalf of that user. Also added simple /me endpoint which will give users/librarians info.